requirements.txt
appdirs==1.4.4
astroid==2.4.2
attrs==20.3.0
beautifulsoup4==4.9.3
behave==1.2.6
black==20.8b1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
config==0.5.0.post0
cryptography==3.2.1
Flask==1.1.2
idna==2.10
iniconfig==1.1.1
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.19.3
packaging==20.8
pandas==1.2.0
parse==1.19.0
parse-type==0.5.2
passlib==1.7.4
pathspec==0.8.1
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycparser==2.20
pydantic==1.7.3
pylint==2.6.0
pyparsing==2.4.7
pytest==6.2.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
PyYAML==5.3.1
redis==3.5.3
regex==2020.11.13
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.25.11
uWSGI==2.0.19.1
Werkzeug==1.0.1
wrapt==1.12.1